refactor: draw in distinct crate
